Графический процесс: добавлен базовый gef.Node
